pyproject.toml
[project]
name = "lunch-recommend-agent"
version = "0.1.0"
description = "Local STT -> LLM -> TTS lunch recommendation agent"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "openai",
  "python-dotenv",
  "gtts",
  "streamlit",
  "sounddevice",
  "scipy",
]
